fixed long long in liblas.hpp, sorry non-gcc users
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40115

+++ include/liblas.hpp
@@ -55,7 +55,7 @@ public:
     void SkipTo(uint64_t nRecord);
 
     /* for writing point records */
-    void SetPoint(const LASPoint &roPoint, long long nPointID);
+    void SetPoint(const LASPoint &roPoint, uint64_t nPointID);
     void AddPoint(const LASPoint &roPoint);
 };
